requirements.txt
mysql-connector-python
fastapi
uvicorn
pydantic
